core/incremental: move the DBSP logical plan next to its only user
The DBSP compiler for materialized views is the only user of the
logical plan in core/translate/logical.rs. Moving the file to
core/incremental frees the translate::logical name for the shared
logical plan stage that the next commits add.

Tests: no behavior change, the file moves as it is.
