+++ core/incremental/mod.rs
@@ -6,6 +6,7 @@ pub mod expr_compiler;
 pub mod filter_operator;
 pub mod input_operator;
 pub mod join_operator;
+pub(crate) mod logical_plan;
 pub mod merge_operator;
 pub mod operator;
 pub mod persistence;
